re-added previously made run/view method
The previous run funciton in the old notebook has ben added in the gui_vis folder. A change was made to make the floor tiles white/gray instead of blue for more clarity. Additionally, the function now returns an Image.Image instance, which makes inline visualizition for notebooks easier with using (dispaly(img)). An xml file can also be saved. this behvaiour could be deleted