small fix mujoco viewer. big progress canonical node
isntead of plt using display for the mujoco viewer, this makes the image larger and better resolution, without a white border. also removed returning the img.

the canonicalnode works 90%. still needs an important bug fix, as it seem to have issues with TOP->TOP and BOTTOM->BOTTOM placements, which can be demonstrated with the stupid.json graph. the canonical_wip notebook contains a way to "semi easily" diy a robot which will be usefull for testing. NOTE: the bug could be related due to the canonicalization of the radial children, not yet tested. The canonicalNode class does not distinguish between chiral parts which are fully plane symmetric (not isomorpic by rotating x degrees). may or may not add this in the future. OPTIMALISATIONS LEFT [Make clockwise/counterclockwise rotation work for axial/radial child order; add lru cache for node class creation in the init; traverse the DFS smarter with doing more actions for each travel, so less overall travel needed; use less for loops for rotations]