progress for important bug
mujoco aligns top and bottom parts from the root in a certain specified way. this means the rotation algorithm requires an extra step if a rotation on the top or bottom was performed. The rotation must happen for the radial parts, that have its axis attached to an axial part through top and bottom. (i.e. root: TOP/BOTTOM -> brick -> something attached to radial brick. if brick rotates, the radial module must also rotate. Needs more research and certaintiy if subtrees are correct. Aditionally, "mental" progress has been made to think of a new algorithm that measures diversity using the subtrees! Idea is to start with the leafs and recreate the newer subtrees from bottom up. turn less-number-of-leaf-tree into larger-number-of-leafs-tree. datastructure and smart algorithm needs to be worked out more in detail