make canonicalnode work closely with mujoco site quats
The CanonicalNode config is being reworked, to make the canonicalnode more accurate. The previous bugs were happening due to the module sites having their own quaternion axis defined. A simple re-rotation was therefore not enough, as the face axis had to also re-rotate. This has been now thoroughly understood and the canonicalNodeConfig is WIP to be reworked to create an adjustment-dictionary. The config makes now use of a newly implemented QuatHolder class that allows for easy arithmatic and comparing operations, as working with np quaternions is very annoying. The config should now accept a set of radial faces and axial faces, with their site quats [which are pre defined and stored in a global dictionary]. using this information, the config can automatically determine the axial_face_order, the shift to re-rotate the list, and the adjustments that still need to calculated. A hardcoded solution is currently implemented in the CanonicalNode class already as a quick fix.