canonicalnode class fully implemented and working
the notebook contains the fully implemented CanonicalNode class which can return the subtrees as strings or nx graphs. The notebook contains a thorough config, with additional insight into quaternion behaviour of the sites. Additionally the contents contain many tests and a profling section. There are still some TODOs, as new inspiration was obtained from researching canonical SMILES and learning about "capturing the morgan neighbourhood" which will reveal even more information than just the subtrees and is also used in chemistry. Aditionally, the repr [the formula used to create the strings] is still to be determined and currently is slightly broken for the CORE.

The contents of this notebook wil be refactored and stored in different places in this repository in the future