added get_xpos
the start of the pointcloud, currently only gathers the 1 point of each component from the simulation. The name is temporary as it will probably be more efficient to make this function return the full point cloud instead of just the xpos of all components.

ps, I put it in view mujoco.py for now because it uses the same imports