NEW: refactoring and documentation of canonical toolkit
the canonical toolkit api has been refactored and integrated from the notebook into the canonical folder. A new grammar method has been created (NEEDS BUG FIXING) to go from strings to graphs and vice versa. The toolkit can be easily used by just importing the CanonicalToolKit. It also contains a new similarity calculation matrics based on the tanimoto metric, but using the counts. It does so for a specific neighbourhood lists given a c ertain radius.

A usage tutorial and can be found in the tutorial.ipynb notebook (WIP).