NEW/WIP: api code refactoring for CanonicalizableNode.
The toolkit for the canonicalizable node felt unintuitive and not nice to use. Because the CanonicalNode contains the children already, it makes more intuitive sense to perform operations on the node already without parsing everything to the toolkit. Inspiration from pandas df has been used for the structure. The toolkit will be useful for initializing the node, or performing operations on 2 nodes together (like analysis). To reduce the size of the class and to create a more clean api, the factory/processor/serializer will still be separate, but use lazy loading and import the class within the method to bypass circular imports (just like pandas). For now this also keeps the toolkit backwards compatible.

Additional changes have been made to the CanonicalizableNode ADT to make the rotation and shifting more user friendly to use, without needing to access internal attributes. The Helper classes should make use of the public api attributes and methods for the CanonicalizableNode, without accessing private attributes.

The ConfigFactory has had changes to remove unused attributes for the CanonicalConfig. Additionally, the radial_shift has been set to default 1 and the face_order has been re-rotated to make this default of 1 work. Additionally some renaming has been done for some attributes. radial_adjustments contains already the precomputed cumulative adjustments to save time.

The config factory still needs to be cleaned.'