[NEW]: factory/node cleanup and new neighbour collection processor (1.5x faster).
Premature optimization is the root of all evil.... Spent too much time on a new collection processor method that doesn't recreate an entire new tree for each radius (per node), but smartly detatched children from the biggest radius. The speedup effect is dissapointing, should have profiled a lot more...... The node api has been restructured a bit and is now more 'clean' and nicely devided into regions. still debating what design choice is good and if the internal classes should be callable from the node itself, instead of the toolkit.
Perhaps the all mutable methods should be applied on the node itself, but the ctk will always return a copy and not mutate the node? Also added stubs

[WIP] The normalisation and tanimoto. more stubs