[NEW] Similarity Calculator with differnt params in config
Evaluator.py has been made to tackle differnt ways on how to calculate the similarity score. This file works close with the toolkit. Exmaple found in the notebook.

The toolkit now contains Enums isntead of literals to pass as 'option' arguments. The toolkit also contains the new similarity metric method that needs a config as argument, the config can be easily created with all the enums in the toolkit, and with calling the create_toolkit method.

TreeDeriver (previously named TreeProcissor) has had a change for the  collect_neighbourhoods method. Previously, it would fill up all nodes with max_radius. If the node maximum was lower, it would just duplicat the highest possible node maximum for the higher radii in the dictionary. I.e. a ' node neigbouurhoud' at radius 7 key, could contain a neighbourhood that given that node, could only make a maximum of 5 hops. Thus, 5,6,7 contained the same neighbourhood for that node. The new use_node_max_radius fixes this. by only going to the max per possible node. (gives more sparse results). tree_max_radius can result in the old behaviour. By not setting this, the max radius will be the absolute maximum hops the tree can make.

The toolkit now also contains an error repressor that is imidately called during import. This works together with the new exceptions.py file found in the utils folder.

When calling get or set item on the node, it will return a custom and 'non invasive' error message instead of the key-error and traceback. Execute node.py to see it in action. Its very nice.

The toolkit can also now call ctk.root or ctk.brick or ctk.hinge instead of ctk.hinge(), as a property, which is nicer to use. to add a rotation, call ctk.create_[hinge/brick/root].

node.py now contains more exported methods (canonicalize (doesnt copy), tograph, tostring) as its just so much easier to use than parsing this into the ctk. The properties are now even more encapsulated, to make it more robust for outside use. The Dataclass wrapper has been removed, but now it contains the visible slots hardcoded. The initializer is now nicer (also for the encapsulation).

tutorial.ipynb has been updated to use the new property calling instead of method calling (i.e. brick instead of brick() ) and it showcases the new similarity score

[TODO] better documentation for the similarity score, and more public ctk methods for ease of use. stubs.